requirements.txt
streamlit
matplotlib
seaborn
plotly
altair
lasio
python-docx
